+++ lib/td_dd_web/views/grant_view.ex
@@ -16,7 +16,16 @@ defmodule TdDdWeb.GrantView do
 
   def render("grant.json", %{grant: grant}) do
     grant
-    |> Map.take([:id, :detail, :start_date, :end_date, :user_id, :source_user_name, :inserted_at, :updated_at])
+    |> Map.take([
+      :id,
+      :detail,
+      :start_date,
+      :end_date,
+      :user_id,
+      :source_user_name,
+      :inserted_at,
+      :updated_at
+    ])
     |> add_structure(grant)
     |> add_structure_version(grant)
     |> add_system(grant)